Changed dataset.
data-set was changed to CASE Dataset by Karan Sharma. Much more information but I have to first re-program data aquisition.
Additionally, the valar data is from a joystick and is not split by different videos. This means I have to split it computationally and calculate an average. This will let me create a model to actually detect the current mood of the person as currently one file contains valar data for multiple different emotions.